push responsibility for creating year dropdown options into sub-component.
